pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "ai-django-core"
authors = [{name = "Ambient Innovation: GmbH", email = "hello@ambient.digital"}]
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.1",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Utilities",
]
dynamic = ["version", "description"]
license = {"file" = "LICENSE.md"}
dependencies = [
    'Django>=2.2.24',
    'bleach>=1.4.3',
    'python-dateutil>=2.5.3',
    'html2text>=2020.1.16',
]

[project.optional-dependencies]
dev = [
    'freezegun',
    'pytest-cov',
    'pytest-django',
    'pytest-mock',
    'gevent',
]
docs = [
    'sphinx==4.2.0',
    'sphinx-rtd-theme==1.0.0',
    'm2r2==0.3.1',
    'mistune<2.0.0', # fixes a problem mit m2r2
]
drf = [
    'djangorestframework>=3.8.2',
]
graphql = [
    'graphene-django>=2.2.0',
    'django-graphql-jwt>=0.2.1',
]
view-layer= [
    'django-crispy-forms>=1.4.0',
]

[tool.flit.module]
name = "ai_django_core"

[project.urls]
'Documentation' = 'https://ai-django-core.readthedocs.io/en/latest/index.html'
'Maintained by' = 'https://ambient.digital/'
'Bugtracker' = 'https://github.com/ambient-innovation/ai-django-core/issues'
'Changelog' = 'https://ai-django-core.readthedocs.io/en/latest/features/changelog.html'

[tool.black]
line-length = 120
target-version = ['py310']
multi_line_output = 3
skip-string-normalization = true
include_trailing_comma = true

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ['ai_django_core', 'testapp', 'tests']

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py{37,38,39,310}-django{22,30,31,32,40}
isolated_build = True

[testenv]
deps =
    django22: Django>=2.2.24,<3.0
    django30: Django>=3.0,<3.1
    django31: Django>=3.1,<3.2
    django32: Django>=3.2,<3.3
    django40: Django>=4.0,<4.1
extras = dev,drf,graphql,view-layer
commands =
    pytest --ds settings tests

[gh-actions]
python =
    3.7: py37
    3.8: py38
    3.9: py39
    3.10: py310
"""
